some changes
